Move identity code
