+++ src/us/kbase/auth2/service/kbase/KBaseAuthConfig.java
@@ -12,10 +12,10 @@ import java.util.Set;
 import org.ini4j.Ini;
 import org.slf4j.LoggerFactory;
 
+import us.kbase.auth2.lib.identity.IdentityProviderConfig;
 import us.kbase.auth2.service.AuthConfig;
 import us.kbase.auth2.service.SLF4JAutoLogger;
 import us.kbase.auth2.service.exceptions.AuthConfigurationException;
-import us.kbase.auth2.service.identity.IdentityProviderConfig;
 import us.kbase.common.service.JsonServerSyslog;
 
 public class KBaseAuthConfig implements AuthConfig {
